Creating Python program run by Raspberry PI 2
This python program is run on the Pi device. It collects the current temperature from the temperature sensor. It also turn the LED on if the temp > 27, else turns it off. It repeatedly sends the current temperature and status of the LED to the MySQL database via a POST http request on a restful API